Refactor existing crypto container and data table. #3
