+++ esp32/main/lib/crypto/data_table.cpp
@@ -1,4 +1,4 @@
-#include "crypto.hpp"
+#include "data_table.hpp"
 
 using namespace Crypto;
